cleaning up workspace table code
