+++ site/src/components/Tooltips/index.ts
@@ -1,0 +1,4 @@
+export { AgentHelpTooltip } from "./AgentHelpTooltip"
+export { OutdatedHelpTooltip } from "./OutdatedHelpTooltip"
+export { ResourcesHelpTooltip } from "./ResourcesHelpTooltip"
+export { WorkspaceHelpTooltip } from "./WorkspaceHelpTooltip"
